Changed return types of these methods after asking Sam

+++ project1/src/main/java/scheduler/Scheduler.java
@@ -20,7 +20,7 @@ public interface Scheduler {
 
     // Statistics
     public int getCurrentBound();
-    public long getTotalStatesVisited();
-    public long getCompleteStatesVisited();
+    public int getTotalStatesVisited();
+    public int getCompleteStatesVisited();
     public int getActiveBranches();
 }
